AI now working as expected
Implemented the mini-max search algorithm with alpha-beta pruning
and a cutoff value to allow imperfect real-time decisions.
The cutoff means a heuristic had to be implemented to estimate the
utility at each state of the board when traversing the game-tree.

The heristic currently takes into account the current score, the
number of corners captured and the mobility of each player.